Windows: Fix 64-bit compile errors (issue #1184).
git-svn-id: https://chromiumembedded.googlecode.com/svn/trunk@1576 5089003a-bbd8-11dd-ad1f-f1f9622dbc98

+++ libcef_dll/wrapper/cef_message_router.cc
@@ -222,9 +222,10 @@ class CefMessageRouterBrowserSideImpl : public CefMessageRouterBrowserSide {
 
       return visitor.count();
     } else if (browser.get()) {
-      return browser_query_info_map_.size(browser->GetIdentifier());
+      return static_cast<int>(
+          browser_query_info_map_.size(browser->GetIdentifier()));
     } else {
-      return browser_query_info_map_.size();
+      return static_cast<int>(browser_query_info_map_.size());
     }
 
     return 0;
@@ -776,9 +777,10 @@ class CefMessageRouterRendererSideImpl : public CefMessageRouterRendererSide {
 
       return visitor.count();
     } else if (browser.get()) {
-      return browser_request_info_map_.size(browser->GetIdentifier());
+      return static_cast<int>(
+          browser_request_info_map_.size(browser->GetIdentifier()));
     } else {
-      return browser_request_info_map_.size();
+      return static_cast<int>(browser_request_info_map_.size());
     }
 
     return 0;
